+++ pyproject.toml
@@ -14,6 +14,7 @@ dependencies = [
     "python-dotenv>=1.2.1",
     "py-cord>=2.6.1",
     "loguru>=0.7.3",
+    "pydantic>=2.12.4",
 ]
 
 [project.scripts]
