+++ src/runescape/api/osrs/__init__.py
@@ -1,0 +1,11 @@
+from enum import Enum
+
+
+class HiscoreType(Enum):
+    NORMAL = "hiscore_oldschool"
+    IRONMAN = "hiscore_oldschool_ironman"
+    ULTIMATE = "hiscore_oldschool_ultimate"
+    HARDCORE = "hiscore_oldschool_hardcore_ironman"
+    DEADMAN = "hiscore_oldschool_deadman"
+    SEASONAL = "hiscore_oldschool_seasonal"
+    TOURNAMENT = "hiscore_oldschool_tournament"
