+++ src/runescape/dataclasses/player.py
@@ -1,0 +1,22 @@
+from pydantic import BaseModel
+
+
+class Skill(BaseModel):
+    id: int
+    name: str
+    rank: int
+    level: int
+    xp: int
+
+
+class Activity(BaseModel):
+    id: int
+    name: str
+    rank: int
+    score: int
+
+
+class Player(BaseModel):
+    name: str
+    skills: list[Skill]
+    activities: list[Activity]
